progs/cursorTest/yt-dlp/requirements.txt
yt-dlp>=2023.11.16
langdetect>=1.0.9
moviepy>=1.0.3
pysrt>=1.1.2
ffmpeg-python>=0.2.0
torch>=2.0.0
openai-whisper>=20231117
transformers>=4.30.0
sentencepiece>=0.1.99
sacremoses>=0.0.53 